language_processor_api/requirements.txt
nltk
gensim
google-cloud-core==0.29.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5